Finished ESP-NOW (sleep not yet enabled)
ESP-NOW is done, and seems to be compatible with the P1.

The data is sent every 32 measurements (with a 10 minute interval
this means every 5 hours). The device then waits for an ACK or a timeout.
If the timeout is reached, it will retry for MAX_SEND_ATTEMPTS-2. If the
maximum number of retries is reached, it will simply abort.

If an ack is received, the device will increase the message index by one,
otherwise it won't.