ESP NOW version with sleep for 10 minutes
ESP NOW now works with light sleep properly (light sleep time was set to 20 seconds instead
of 10 minutes, is fixed now).

When you want to test this, you will have to change the MAC address of
the sender and receiver, to the MAC adresses of your devices.

In this commit, the device will measure CO2, RH and temperature every ten minutes. In between
it will be in light sleep. When there have been 32 measurements, the device will try to
send the data (it will disable modem sleep). If the device
did not receive an ACK to its send, it will retry for 0xFFFF times (see MAX_SEND_ATTEMPTS)
in espnow.c. After these attempts, it will simply abort.

If the device was succesfull - or aborted - in sending its message, it will go
back to light sleep.